member dashboard component로 분리
